// src/shared/api/generated/models/examinationManageContractDtoUserUserLoginInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */
import type { ExaminationManageContractDtoUserUserLoginInputSystem } from './examinationManageContractDtoUserUserLoginInputSystem';
import type { ExaminationManageContractDtoUserUserLoginInputClientType } from './examinationManageContractDtoUserUserLoginInputClientType';

export interface ExaminationManageContractDtoUserUserLoginInput {
  code?: string;
  system?: ExaminationManageContractDtoUserUserLoginInputSystem;
  /** 省 */
  province?: string;
  /** 省 */
  provinceCode?: string;
  /** 科目Id */
  examTypeId?: string;
  /** 输入的邀请码 */
  iptInviteCode?: string;
  terminal?: string;
  clientType?: ExaminationManageContractDtoUserUserLoginInputClientType;
  package?: string;
}