// src/shared/api/generated/models/examinationManageContractDtoUserShortMessageLoginInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */
import type { ExaminationManageContractDtoUserShortMessageLoginInputSystem } from './examinationManageContractDtoUserShortMessageLoginInputSystem';
import type { ExaminationManageContractDtoUserShortMessageLoginInputClientType } from './examinationManageContractDtoUserShortMessageLoginInputClientType';

export interface ExaminationManageContractDtoUserShortMessageLoginInput {
  mobile?: string;
  code?: string;
  requestId?: string;
  system?: ExaminationManageContractDtoUserShortMessageLoginInputSystem;
  /** 省 */
  province?: string;
  /** 省Code */
  provinceCode?: string;
  /** 科目Id */
  examTypeId?: string;
  /** 头像 */
  headImg?: string;
  /** 输入的邀请码 */
  iptInviteCode?: string;
  terminal?: string;
  clientType?: ExaminationManageContractDtoUserShortMessageLoginInputClientType;
}